[NB] Remove diffuser, replace with new camera-h...
Remove light-diffuser box, replace with new camera-holder, mounted using side
arms, rather than acrylic side-plates.

__NB__ The new camera-holder has a flat face where a light-source _(e.g., LED
array)_ may be mounted.  The top of the camera-holder has extruded plates on
either side, providing a place to attach a piece of paper using clips.  The
piece of paper can then be draped over the device, with the opposite end of the
piece of paper tucked under the front of the base plate, such that the paper
diffuses the light from the upward-facing light-source down to the device.